integrated pose estimation in microservice and updated colors

+++ flask_wrapper/code_b/angles.py
@@ -14,7 +14,7 @@ def calc_angle(landmark1, landmark2):
 
 
 def angle_hip(hip_l, hip_r):
-    hip_v = np.array([hip_l.x - hip_r.x, hip_l.y - hip_r.y, hip_l.z - hip_r.z], dtype=np.float16)
+    hip_v = np.array([hip_l.x - hip_r.x, hip_l.y - hip_r.y, hip_l.z - hip_r.z], dtype=np.float64)
     normal = np.array([0, 0, 1])
     angle = np.arccos(normal.dot(hip_v) / (np.linalg.norm(normal) * np.linalg.norm(hip_v)))
     angle = 90 - np.degrees(angle)
@@ -23,7 +23,7 @@ def angle_hip(hip_l, hip_r):
 
 
 def angle_ground(left, right):
-    vector = np.array([left.x - right.x, left.y - right.y, left.z - right.z], dtype=np.float16)
+    vector = np.array([left.x - right.x, left.y - right.y, left.z - right.z], dtype=np.float64)
     normal = np.array([0, 1, 0])
     angle = np.arccos(normal.dot(vector) / (np.linalg.norm(normal) * np.linalg.norm(vector)))
     angle = 90 - np.degrees(angle)
@@ -32,8 +32,8 @@ def angle_ground(left, right):
 
 
 def back_angle(shoulder_l, shoulder_r):
-    sl = np.array([shoulder_l.x, shoulder_l.y, shoulder_l.z], dtype=np.float16)
-    sr = np.array([shoulder_r.x, shoulder_r.y, shoulder_r.z], dtype=np.float16)
+    sl = np.array([shoulder_l.x, shoulder_l.y, shoulder_l.z], dtype=np.float64)
+    sr = np.array([shoulder_r.x, shoulder_r.y, shoulder_r.z], dtype=np.float64)
     connection = sl - sr
     spine = sl + 0.5 * connection
     normal = np.array([0, 1, 0])
@@ -44,8 +44,8 @@ def back_angle(shoulder_l, shoulder_r):
 
 
 def tilt_angle(shoulder_l, shoulder_r):
-    sl = np.array([shoulder_l.x, shoulder_l.y, shoulder_l.z], dtype=np.float16)
-    sr = np.array([shoulder_r.x, shoulder_r.y, shoulder_r.z], dtype=np.float16)
+    sl = np.array([shoulder_l.x, shoulder_l.y, shoulder_l.z], dtype=np.float64)
+    sr = np.array([shoulder_r.x, shoulder_r.y, shoulder_r.z], dtype=np.float64)
     connection = sl - sr
     spine = sl + 0.5 * connection
     normal = np.array([0, 1, 0])
@@ -56,8 +56,8 @@ def tilt_angle(shoulder_l, shoulder_r):
 
 
 def mass_balance(foot_l, foot_r):
-    left = np.array([foot_l.x, foot_l.y, foot_l.z], dtype=np.float16)
-    right = np.array([foot_r.x, foot_r.y, foot_r.z], dtype=np.float16)
+    left = np.array([foot_l.x, foot_l.y, foot_l.z], dtype=np.float64)
+    right = np.array([foot_r.x, foot_r.y, foot_r.z], dtype=np.float64)
     distance_l = np.linalg.norm(left)
     distance_r = np.linalg.norm(right)
 
@@ -74,7 +74,7 @@ def mass_balance(foot_l, foot_r):
 
 
 def pelvis_rotation(hip_l, hip_r, r):
-    hip_v = r @ np.array([hip_l.x - hip_r.x, 0, hip_l.z - hip_r.z], dtype=np.float16)
+    hip_v = r @ np.array([hip_l.x - hip_r.x, 0, hip_l.z - hip_r.z], dtype=np.float64)
     normal = np.array([0, 0, 1])
     # angle = np.arccos(normal.dot(hip_v) / (np.linalg.norm(normal) * np.linalg.norm(hip_v)))
     # angle between hip vector and normal with atan2
@@ -84,7 +84,7 @@ def pelvis_rotation(hip_l, hip_r, r):
 
 
 def pelvis_tilt(hip_l, hip_r, r):
-    hip_v = r @ np.array([hip_l.x - hip_r.x, hip_l.y - hip_r.y, hip_l.z - hip_r.z], dtype=np.float16)
+    hip_v = r @ np.array([hip_l.x - hip_r.x, hip_l.y - hip_r.y, hip_l.z - hip_r.z], dtype=np.float64)
     normal = np.array([0, 1, 0])
     angle = np.arccos(normal.dot(hip_v) / (np.linalg.norm(normal) * np.linalg.norm(hip_v)))
 
@@ -92,25 +92,25 @@ def pelvis_tilt(hip_l, hip_r, r):
 
 
 def pelvis_sway(foot_l, r):
-    hip_v = r @ np.array([foot_l.x, foot_l.y, foot_l.z], dtype=np.float16)
+    hip_v = r @ np.array([foot_l.x, foot_l.y, foot_l.z], dtype=np.float64)
 
     return hip_v[2]
 
 
 def pelvis_thrust(foot_l, r):
-    hip_v = r @ np.array([foot_l.x, foot_l.y, foot_l.z], dtype=np.float16)
+    hip_v = r @ np.array([foot_l.x, foot_l.y, foot_l.z], dtype=np.float64)
 
     return hip_v[0]
 
 
 def pelvis_lift(foot_l, r):
-    hip_v = r @ np.array([foot_l.x, foot_l.y, foot_l.z], dtype=np.float16)
+    hip_v = r @ np.array([foot_l.x, foot_l.y, foot_l.z], dtype=np.float64)
 
     return hip_v[1]
 
 
 def thorax_rotation(shoulder_l, shoulder_r, r):
-    shoulder_v = r @ np.array([shoulder_l.x - shoulder_r.x, 0, shoulder_l.z - shoulder_r.z], dtype=np.float16)
+    shoulder_v = r @ np.array([shoulder_l.x - shoulder_r.x, 0, shoulder_l.z - shoulder_r.z], dtype=np.float64)
     normal = np.array([0, 0, 1])
     # angle = np.arccos(normal.dot(shoulder_v) / (np.linalg.norm(normal) * np.linalg.norm(shoulder_v)))
     # angle between shoulder vector and normal with atan2
@@ -120,9 +120,9 @@ def thorax_rotation(shoulder_l, shoulder_r, r):
 
 
 def thorax_tilt(shoulder_l, shoulder_r, r):
-    shoulder_v = r @ np.array([shoulder_l.x - shoulder_r.x, shoulder_l.y - shoulder_r.y, shoulder_l.z - shoulder_r.z], dtype=np.float16)
+    shoulder_v = r @ np.array([shoulder_l.x - shoulder_r.x, shoulder_l.y - shoulder_r.y, shoulder_l.z - shoulder_r.z], dtype=np.float64)
     shoulder_v[0] = 0
-    # shoulder_l = r @ np.array([shoulder_l.x, shoulder_l.y, shoulder_l.z], dtype=np.float16)
+    # shoulder_l = r @ np.array([shoulder_l.x, shoulder_l.y, shoulder_l.z], dtype=np.float64)
     # spine = shoulder_l - 0.5 * shoulder_v
     # spine[0] = 0
     normal = np.array([0, 1, 0])
@@ -135,8 +135,8 @@ def thorax_tilt(shoulder_l, shoulder_r, r):
 
 
 def thorax_bend(shoulder_l, shoulder_r, r):
-    shoulder_v = r @ np.array([shoulder_l.x - shoulder_r.x, shoulder_l.y - shoulder_r.y, shoulder_l.z - shoulder_r.z], dtype=np.float16)
-    shoulder_l = r @ np.array([shoulder_l.x, shoulder_l.y, shoulder_l.z], dtype=np.float16)
+    shoulder_v = r @ np.array([shoulder_l.x - shoulder_r.x, shoulder_l.y - shoulder_r.y, shoulder_l.z - shoulder_r.z], dtype=np.float64)
+    shoulder_l = r @ np.array([shoulder_l.x, shoulder_l.y, shoulder_l.z], dtype=np.float64)
     spine = shoulder_l - 0.5 * shoulder_v
     spine[2] = 0
     normal = np.array([0, 1, 0])
@@ -146,9 +146,9 @@ def thorax_bend(shoulder_l, shoulder_r, r):
 
 
 def thorax_sway(shoulder_l, shoulder_r, foot_l, r):
-    shoulder_v = r @ np.array([shoulder_l.x - shoulder_r.x, shoulder_l.y - shoulder_r.y, shoulder_l.z - shoulder_r.z], dtype=np.float16)
-    shoulder_l = r @ np.array([shoulder_l.x, shoulder_l.y, shoulder_l.z], dtype=np.float16)
-    foot_l = r @ np.array([foot_l.x, foot_l.y, foot_l.z], dtype=np.float16)
+    shoulder_v = r @ np.array([shoulder_l.x - shoulder_r.x, shoulder_l.y - shoulder_r.y, shoulder_l.z - shoulder_r.z], dtype=np.float64)
+    shoulder_l = r @ np.array([shoulder_l.x, shoulder_l.y, shoulder_l.z], dtype=np.float64)
+    foot_l = r @ np.array([foot_l.x, foot_l.y, foot_l.z], dtype=np.float64)
     spine = shoulder_l - 0.5 * shoulder_v
     sway = spine - foot_l
 
@@ -156,9 +156,9 @@ def thorax_sway(shoulder_l, shoulder_r, foot_l, r):
 
 
 def thorax_thrust(shoulder_l, shoulder_r, foot_l, r):
-    shoulder_v = r @ np.array([shoulder_l.x - shoulder_r.x, shoulder_l.y - shoulder_r.y, shoulder_l.z - shoulder_r.z], dtype=np.float16)
-    shoulder_l = r @ np.array([shoulder_l.x, shoulder_l.y, shoulder_l.z], dtype=np.float16)
-    foot_l = r @ np.array([foot_l.x, foot_l.x, foot_l.x], dtype=np.float16)
+    shoulder_v = r @ np.array([shoulder_l.x - shoulder_r.x, shoulder_l.y - shoulder_r.y, shoulder_l.z - shoulder_r.z], dtype=np.float64)
+    shoulder_l = r @ np.array([shoulder_l.x, shoulder_l.y, shoulder_l.z], dtype=np.float64)
+    foot_l = r @ np.array([foot_l.x, foot_l.x, foot_l.x], dtype=np.float64)
     spine = shoulder_l - 0.5 * shoulder_v
     thrust = spine - foot_l
 
@@ -166,9 +166,9 @@ def thorax_thrust(shoulder_l, shoulder_r, foot_l, r):
 
 
 def thorax_lift(shoulder_l, shoulder_r, foot_l, r):
-    shoulder_v = r @ np.array([shoulder_l.x - shoulder_r.x, shoulder_l.y - shoulder_r.y, shoulder_l.z - shoulder_r.z], dtype=np.float16)
-    shoulder_l = r @ np.array([shoulder_l.x, shoulder_l.y, shoulder_l.z], dtype=np.float16)
-    foot_l = r @ np.array([foot_l.x, foot_l.y, foot_l.z], dtype=np.float16)
+    shoulder_v = r @ np.array([shoulder_l.x - shoulder_r.x, shoulder_l.y - shoulder_r.y, shoulder_l.z - shoulder_r.z], dtype=np.float64)
+    shoulder_l = r @ np.array([shoulder_l.x, shoulder_l.y, shoulder_l.z], dtype=np.float64)
+    foot_l = r @ np.array([foot_l.x, foot_l.y, foot_l.z], dtype=np.float64)
     spine = shoulder_l - 0.5 * shoulder_v
     lift = spine - foot_l
 
@@ -177,8 +177,8 @@ def thorax_lift(shoulder_l, shoulder_r, foot_l, r):
 
 # rotation between spine and pelvis
 def spine_rotation(hip_l, hip_r, shoulder_l, shoulder_r, r):
-    hip_v = r @ np.array([hip_l.x - hip_r.x, 0, hip_l.z - hip_r.z], dtype=np.float16)
-    shoulder_v = r @ np.array([shoulder_l.x - shoulder_r.x, 0, shoulder_l.z - shoulder_r.z], dtype=np.float16)
+    hip_v = r @ np.array([hip_l.x - hip_r.x, 0, hip_l.z - hip_r.z], dtype=np.float64)
+    shoulder_v = r @ np.array([shoulder_l.x - shoulder_r.x, 0, shoulder_l.z - shoulder_r.z], dtype=np.float64)
     # angle = np.arccos(hip_v.dot(shoulder_v) / (np.linalg.norm(hip_v) * np.linalg.norm(shoulder_v)))
     # angle between hip vector and shoulder vector with atan2
     angle = (np.arctan2(hip_v[2] * shoulder_v[0] - hip_v[0] * shoulder_v[2],
@@ -189,8 +189,8 @@ def spine_rotation(hip_l, hip_r, shoulder_l, shoulder_r, r):
 
 # tilt between spine and pelvis
 def spine_tilt(hip_l, hip_r, shoulder_l, shoulder_r, r):
-    hip_v = r @ np.array([hip_l.x - hip_r.x, hip_l.y - hip_r.y, hip_l.z - hip_r.z], dtype=np.float16)
-    shoulder_v = r @ np.array([shoulder_l.x - shoulder_r.x, shoulder_l.y - shoulder_r.y, shoulder_l.z - shoulder_r.z], dtype=np.float16)
+    hip_v = r @ np.array([hip_l.x - hip_r.x, hip_l.y - hip_r.y, hip_l.z - hip_r.z], dtype=np.float64)
+    shoulder_v = r @ np.array([shoulder_l.x - shoulder_r.x, shoulder_l.y - shoulder_r.y, shoulder_l.z - shoulder_r.z], dtype=np.float64)
     angle = np.arccos(hip_v.dot(shoulder_v) / (np.linalg.norm(hip_v) * np.linalg.norm(shoulder_v)))
 
     return np.degrees(angle)
@@ -198,7 +198,7 @@ def spine_tilt(hip_l, hip_r, shoulder_l, shoulder_r, r):
 
 # head rotation
 def head_rotation(eye_l, eye_r, r):
-    eye_v = r @ np.array([eye_l.x - eye_r.x, 0, eye_l.z - eye_r.z], dtype=np.float16)
+    eye_v = r @ np.array([eye_l.x - eye_r.x, 0, eye_l.z - eye_r.z], dtype=np.float64)
     normal = np.array([0, 0, 1])
     # angle = np.arccos(normal.dot(eye_v) / (np.linalg.norm(normal) * np.linalg.norm(eye_v)))
     # angle between eye vector and normal with atan2
@@ -209,7 +209,7 @@ def head_rotation(eye_l, eye_r, r):
 
 # head tilt
 def head_tilt(eye_l, eye_r, r):
-    eye_v = r @ np.array([eye_l.x - eye_r.x, eye_l.y - eye_r.y, eye_l.z - eye_r.z], dtype=np.float16)
+    eye_v = r @ np.array([eye_l.x - eye_r.x, eye_l.y - eye_r.y, eye_l.z - eye_r.z], dtype=np.float64)
     eye_v[0] = 0
     normal = np.array([0, 0, 1])
     angle = np.arccos(normal.dot(eye_v) / (np.linalg.norm(normal) * np.linalg.norm(eye_v)))
@@ -219,9 +219,9 @@ def head_tilt(eye_l, eye_r, r):
 
 # distance from spine to left wrist
 def left_arm_length(shoulder_l, shoulder_r, wrist_l, r):
-    shoulder_v = r @ np.array([shoulder_l.x - shoulder_r.x, shoulder_l.y - shoulder_r.y, shoulder_l.z - shoulder_r.z], dtype=np.float16)
-    shoulder_l = r @ np.array([shoulder_l.x, shoulder_l.y, shoulder_l.z], dtype=np.float16)
-    wrist_l = r @ np.array([wrist_l.x, wrist_l.y, wrist_l.z], dtype=np.float16)
+    shoulder_v = r @ np.array([shoulder_l.x - shoulder_r.x, shoulder_l.y - shoulder_r.y, shoulder_l.z - shoulder_r.z], dtype=np.float64)
+    shoulder_l = r @ np.array([shoulder_l.x, shoulder_l.y, shoulder_l.z], dtype=np.float64)
+    wrist_l = r @ np.array([wrist_l.x, wrist_l.y, wrist_l.z], dtype=np.float64)
     spine = shoulder_l - 0.5 * shoulder_v
     length = np.linalg.norm(spine - wrist_l)
 
@@ -229,10 +229,10 @@ def left_arm_length(shoulder_l, shoulder_r, wrist_l, r):
 
 
 def wrist_angle(pinky_l, index_l, wrist_l, elbow_l, r):
-    pinky_l = r @ np.array([pinky_l.x, pinky_l.y, pinky_l.z], dtype=np.float16)
-    index_l = r @ np.array([index_l.x, index_l.y, index_l.z], dtype=np.float16)
-    wrist_l = r @ np.array([wrist_l.x, wrist_l.y, wrist_l.z], dtype=np.float16)
-    elbow_l = r @ np.array([elbow_l.x, elbow_l.y, elbow_l.z], dtype=np.float16)
+    pinky_l = r @ np.array([pinky_l.x, pinky_l.y, pinky_l.z], dtype=np.float64)
+    index_l = r @ np.array([index_l.x, index_l.y, index_l.z], dtype=np.float64)
+    wrist_l = r @ np.array([wrist_l.x, wrist_l.y, wrist_l.z], dtype=np.float64)
+    elbow_l = r @ np.array([elbow_l.x, elbow_l.y, elbow_l.z], dtype=np.float64)
     pinky_v = pinky_l - wrist_l
     index_v = index_l - wrist_l
     normal = np.cross(pinky_v, index_v)
@@ -245,10 +245,10 @@ def wrist_angle(pinky_l, index_l, wrist_l, elbow_l, r):
     return np.degrees(angle) + 90
 
 def wrist_tilt(pinky_l, index_l, wrist_l, elbow_l, r):
-    pinky_l = r @ np.array([pinky_l.x, pinky_l.y, pinky_l.z], dtype=np.float16)
-    index_l = r @ np.array([index_l.x, index_l.y, index_l.z], dtype=np.float16)
-    wrist_l = r @ np.array([wrist_l.x, wrist_l.y, wrist_l.z], dtype=np.float16)
-    elbow_l = r @ np.array([elbow_l.x, elbow_l.y, elbow_l.z], dtype=np.float16)
+    pinky_l = r @ np.array([pinky_l.x, pinky_l.y, pinky_l.z], dtype=np.float64)
+    index_l = r @ np.array([index_l.x, index_l.y, index_l.z], dtype=np.float64)
+    wrist_l = r @ np.array([wrist_l.x, wrist_l.y, wrist_l.z], dtype=np.float64)
+    elbow_l = r @ np.array([elbow_l.x, elbow_l.y, elbow_l.z], dtype=np.float64)
     hand_v = pinky_l - index_l
     arm_l = wrist_l - elbow_l
     # angle with atan2
@@ -259,7 +259,7 @@ def wrist_tilt(pinky_l, index_l, wrist_l, elbow_l, r):
 
 
 def arm_rotation(wrist_l, r):
-    wrist_v = r @ np.array([wrist_l.x, wrist_l.y, wrist_l.z], dtype=np.float16)
+    wrist_v = r @ np.array([wrist_l.x, wrist_l.y, wrist_l.z], dtype=np.float64)
     wrist_v[1] = 0
     normal = np.array([1, 0, 0])
     # angle = np.arccos(normal.dot(wrist_v) / (np.linalg.norm(normal) * np.linalg.norm(wrist_v)))
